requirements.txt
yfinance
pytest
pandas
numpy
matplotlib
seaborn
pyyaml
pyfiglet
yahoo_earnings_calendar
urllib3
beautifulsoup4
html5lib
plotly
dash
dash-bootstrap-components
